+++ packages/categories/i18n/en.js
@@ -2,9 +2,10 @@
 i18n.map('en', {
 	title_cats_places: "Places by Category: {$1}",
 	title_cats_users: "Users by Category: {$1}",
-	title_panelCatsType: "Manage Categories",
+	title_panelAdminCatsType: "Manage Categories",
 	label_cat: "Category",
 	label_cats: "Categories",
+	label_catsall: "All Categories",
 	label_catshist: "Last used categories",
 	error_cats_exists: "Category {$1} just exists"
 });